+++ packages/dhis2/src/Adaptor.js
@@ -787,6 +787,7 @@ export {
   fn,
   fnIf,
   http,
+  group,
   lastReferenceValue,
   merge,
   sourceValue,