+++ packages/kobotoolbox/src/Adaptor.js
@@ -161,6 +161,7 @@ export {
   fn,
   fnIf,
   http,
+  group,
   lastReferenceValue,
   merge,
   sourceValue,