+++ packages/http/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @openfn/language-http
 
+## 6.4.0
+
+### Minor Changes
+
+- 5fb82f07: Export `group` operation from common
+
+### Patch Changes
+
+- Updated dependencies [5fb82f07]
+  - @openfn/language-common@1.15.0
+
 ## 6.3.0
 
 ### Minor Changes